requirements.txt
accelerate >= 0.12.0
datasets >= 1.8.0
sentencepiece != 0.1.92
scipy
scikit-learn
protobuf
torch >= 1.3
evaluate
transformers
streamlit
faiss-cpu
sentence-transformers